docs/requirements.txt
Sphinx >= 3.3, < 3.4
docutils >= 0.16, < 0.17
sphinx-better-theme >= 0.1.5, < 0.2
